refactor(config): isolate runtime auth policy interpretation
